Add userID to category & todo(delete/edit) routes

+++ routes/delete-todo.js
@@ -6,10 +6,12 @@ module.exports = (db) => {
   //Queries DELETE todo
   router.get("/:id", (req, res) => {
     const todoID = req.params.id;
-    const queryParams = [todoID];
+    const userID = req.session.user_id;
+    const queryParams = [todoID, userID];
     const queryString = `
     DELETE FROM todos
     WHERE id = $1
+    AND user_id = $2
     RETURNING *;`;
 
     db.query(queryString, queryParams)
